requirements.txt
pandas
openpyxl
fastapi
uvicorn
chromadb
sentence-transformers
rank-bm25
pypdf
python-docx
beautifulsoup4
ollama